requirements.txt
PyQt5>=5.15.0
requests>=2.31.0
qrcode>=7.4.0
Pillow>=10.0.0
yt-dlp>=2023.11.0
ffmpeg-python>=0.2.0
